[BY-255] Move: make folder at widget/auth/setup
